Working again

+++ include/cbmdefines.h
@@ -50,8 +50,8 @@ enum IECChannels {
 #define CBM_EXCLAMATION_MARKS "!!"
 #define CBM_DOLLAR_SIGN '$'
 
-const int MAX_CBM_SCREEN_ROWS = 25;
-const int MAX_CBM_SCREEN_COLS = 40;
+const uint8_t MAX_CBM_SCREEN_ROWS = 25;
+const uint8_t MAX_CBM_SCREEN_COLS = 40;
 
 ////////////////////////////////////////////////////////////////////////////////
 //
