+++ src/ESPWebDAV/ESPWebDAV.h
@@ -19,7 +19,7 @@ enum DepthType { DEPTH_NONE, DEPTH_CHILD, DEPTH_ALL };
 
 class ESPWebDAV	{
 public:
-	bool init(int serverPort, FS* fileSystem);
+	bool init(uint8_t serverPort, FS* fileSystem);
 	bool stop();
 	bool isClientWaiting();
 	void handleClient(String blank = "");
@@ -45,7 +45,7 @@ protected:
 	void handleMove(ResourceType resource);
 	void handleDelete(ResourceType resource);
 
-	int deleteRecursive(String path);
+	uint8_t deleteRecursive(String path);
 
 	// Sections are copied from ESP8266Webserver
 	String getMimeType(String path);
