+++ src/ESPWebDAV/WebSrv.cpp
@@ -39,8 +39,8 @@ String ESPWebDAV::urlDecode(const String& text)	{
 // ------------------------
 	String decoded = "";
 	char temp[] = "0x00";
-	unsigned int len = text.length();
-	unsigned int i = 0;
+	uint8_t len = text.length();
+	uint8_t i = 0;
 	while (i < len)	{
 		char decodedChar;
 		char encodedChar = text.charAt(i++);
@@ -68,7 +68,7 @@ String ESPWebDAV::urlDecode(const String& text)	{
 String ESPWebDAV::urlToUri(String url)	{
 // ------------------------
 	if(url.startsWith("http://"))	{
-		int uriStart = url.indexOf('/', 7);
+		uint8_t uriStart = url.indexOf('/', 7);
 		return url.substring(uriStart);
 	}
 	else
@@ -153,8 +153,8 @@ bool ESPWebDAV::parseRequest() {
 	
 	// First line of HTTP request looks like "GET /path HTTP/1.1"
 	// Retrieve the "/path" part by finding the spaces
-	int addr_start = req.indexOf(' ');
-	int addr_end = req.indexOf(' ', addr_start + 1);
+	uint8_t addr_start = req.indexOf(' ');
+	uint8_t addr_end = req.indexOf(' ', addr_start + 1);
 	if (addr_start == -1 || addr_end == -1) {
 		return false;
 	}
@@ -174,7 +174,7 @@ bool ESPWebDAV::parseRequest() {
 			// no more headers
 			break;
 			
-		int headerDiv = req.indexOf(':');
+		uint8_t headerDiv = req.indexOf(':');
 		if (headerDiv == -1)
 			break;
 		
@@ -313,7 +313,7 @@ void ESPWebDAV::setContentLength(size_t len)	{
 // ------------------------
 size_t ESPWebDAV::readBytesWithTimeout(uint8_t *buf, size_t bufSize) {
 // ------------------------
-	int timeout_ms = HTTP_MAX_POST_WAIT;
+	uint8_t timeout_ms = HTTP_MAX_POST_WAIT;
 	size_t numAvailable = 0;
 	while(!(numAvailable = client.available()) && client.connected() && timeout_ms--) 
 		delay(1);
@@ -328,7 +328,7 @@ size_t ESPWebDAV::readBytesWithTimeout(uint8_t *buf, size_t bufSize) {
 // ------------------------
 size_t ESPWebDAV::readBytesWithTimeout(uint8_t *buf, size_t bufSize, size_t numToRead) {
 // ------------------------
-	int timeout_ms = HTTP_MAX_POST_WAIT;
+	uint8_t timeout_ms = HTTP_MAX_POST_WAIT;
 	size_t numAvailable = 0;
 	
 	while(((numAvailable = client.available()) < numToRead) && client.connected() && timeout_ms--)
